Move pyani_plus.snakemake.snakemake_scheduler -> pyani_plus.workflows
No longer need to use importlib.resources to work out where
the snakemake files are - they are along side the Python
file in the same directory.
