+++ pyani_plus/workflows/__init__.py
@@ -19,4 +19,69 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
-"""Module providing snakemake workflows."""
+"""Snakemake workflows for ANI pairwise comparisons."""
+
+from pathlib import Path
+
+from snakemake.api import ConfigSettings, DAGSettings, ResourceSettings, SnakemakeApi
+from snakemake.settings.types import OutputSettings, Quietness
+
+
+def check_input_stems(indir: str) -> dict[str, Path]:
+    """Check input files against approved list of extensions.
+
+    If duplicate stems with approved extensions are present
+    raise a ValueError.
+    """
+    extensions = [".fasta", ".fas", ".fna"]
+    stems = [_.stem for _ in Path(indir).glob("*") if _.suffix in extensions]
+
+    if len(stems) == len(set(stems)):
+        input_files = {
+            _.stem: _ for _ in Path(indir).glob("*") if _.suffix in extensions
+        }
+    else:
+        duplicates = [
+            item for item in stems if stems.count(item) > 1 and item in set(stems)
+        ]
+        msg = (
+            f"Duplicated stems found for {sorted(set(duplicates))}. Please investigate."
+        )
+        raise ValueError(msg)
+
+    return input_files
+
+
+class SnakemakeRunner:
+    """Execute Snakemake workflows for pairwise comparisons."""
+
+    def __init__(self, workflow_filename: str) -> None:
+        """Initialise SnakemakeRunner with SnakeMake workflow filename."""
+        self.workflow_filename = workflow_filename
+
+    def run_workflow(
+        self, targetset: list, config_args: dict, workdir: Path | None = None
+    ) -> None:
+        """Run the snakemake workflow.
+
+        :param targets: (list) List of target files that the workflow should return.
+        :config_args: (dict) Dictionary of configuration arguments to pass to the smk workflow.
+        """
+        targetset = [str(_) for _ in targetset]
+
+        # Path to snakemake workflow file, which will be next to this python file
+        snakefile = Path(__file__).with_name(self.workflow_filename)
+
+        # Use the defined workflow from the Python API
+        with SnakemakeApi(OutputSettings(quiet={Quietness.ALL})) as snakemake_api:
+            workflow_api = snakemake_api.workflow(
+                snakefile=snakefile,
+                resource_settings=ResourceSettings(cores=config_args["cores"]),
+                config_settings=ConfigSettings(config=config_args),
+                workdir=workdir,
+            )
+            dag_api = workflow_api.dag(
+                dag_settings=DAGSettings(targets=targetset),
+            )
+            dag_api.unlock()
+            dag_api.execute_workflow()
