+++ tests/snakemake/test_snakemake_anib_workflow.py
@@ -34,8 +34,8 @@ from pathlib import Path
 import pytest
 
 from pyani_plus.private_cli import log_configuration, log_genome, log_run
-from pyani_plus.snakemake import snakemake_scheduler
 from pyani_plus.tools import get_blastn, get_makeblastdb
+from pyani_plus.workflows import SnakemakeRunner, check_input_stems
 
 from . import compare_matrices
 
@@ -75,7 +75,7 @@ def test_snakemake_rule_fragments(
     shutil.rmtree(anib_targets_outdir, ignore_errors=True)
 
     # Run snakemake wrapper
-    runner = snakemake_scheduler.SnakemakeRunner("snakemake_anib.smk")
+    runner = SnakemakeRunner("snakemake_anib.smk")
     runner.run_workflow(
         anib_targets_fragments, config_anib_args, workdir=Path(tmp_path)
     )
@@ -112,7 +112,7 @@ def test_snakemake_rule_blastdb(
     shutil.rmtree(anib_targets_outdir, ignore_errors=True)
 
     # Run snakemake wrapper
-    runner = snakemake_scheduler.SnakemakeRunner("snakemake_anib.smk")
+    runner = SnakemakeRunner("snakemake_anib.smk")
     runner.run_workflow(anib_targets_blastdb, config_anib_args, workdir=Path(tmp_path))
 
     # Check output against target fixtures
@@ -151,14 +151,12 @@ def test_snakemake_rule_blastn(  # noqa: PLR0913
     # Record the FASTA files in the genomes table _before_ call snakemake
     log_genome(
         database=db,
-        fasta=list(
-            snakemake_scheduler.check_input_stems(config_anib_args["indir"]).values()
-        ),
+        fasta=list(check_input_stems(config_anib_args["indir"]).values()),
     )
     assert db.is_file()
 
     # Run snakemake wrapper
-    runner = snakemake_scheduler.SnakemakeRunner("snakemake_anib.smk")
+    runner = SnakemakeRunner("snakemake_anib.smk")
     runner.run_workflow(anib_targets_blastn, config_anib_args, workdir=Path(tmp_path))
 
     # Check output against target fixtures
