+++ tests/snakemake/test_snakemake_dnadiff_workflow.py
@@ -32,13 +32,13 @@ from pathlib import Path
 import pytest
 
 from pyani_plus.private_cli import log_configuration, log_genome, log_run
-from pyani_plus.snakemake import snakemake_scheduler
 from pyani_plus.tools import (
     get_delta_filter,
     get_nucmer,
     get_show_coords,
     get_show_diff,
 )
+from pyani_plus.workflows import SnakemakeRunner, check_input_stems
 
 from . import compare_matrices
 
@@ -123,7 +123,7 @@ def test_snakemake_rule_delta(
     config["outdir"] = dnadiff_nucmer_targets_delta_outdir
 
     # Run snakemake wrapper
-    runner = snakemake_scheduler.SnakemakeRunner("snakemake_dnadiff.smk")
+    runner = SnakemakeRunner("snakemake_dnadiff.smk")
     runner.run_workflow(dnadiff_nucmer_targets_delta, config, workdir=Path(tmp_path))
 
     # Check output against target fixtures
@@ -159,7 +159,7 @@ def test_snakemake_rule_filter(
     config["outdir"] = dnadiff_nucmer_targets_filter_outdir
 
     # Run snakemake wrapper
-    runner = snakemake_scheduler.SnakemakeRunner("snakemake_dnadiff.smk")
+    runner = SnakemakeRunner("snakemake_dnadiff.smk")
     runner.run_workflow(dnadiff_nucmer_targets_filter, config, workdir=Path(tmp_path))
 
     # Check output against target fixtures
@@ -212,9 +212,7 @@ def test_snakemake_rule_show_diff_and_coords(  # noqa: PLR0913
     # Record the FASTA files in the genomes table _before_ call snakemake
     log_genome(
         database=db,
-        fasta=list(
-            snakemake_scheduler.check_input_stems(config_dnadiff_args["indir"]).values()
-        ),
+        fasta=list(check_input_stems(config_dnadiff_args["indir"]).values()),
     )
     assert db.is_file()
 
@@ -222,7 +220,7 @@ def test_snakemake_rule_show_diff_and_coords(  # noqa: PLR0913
     config["outdir"] = dnadiff_targets_showdiff_outdir
 
     # Run snakemake wrapper
-    runner = snakemake_scheduler.SnakemakeRunner("snakemake_dnadiff.smk")
+    runner = SnakemakeRunner("snakemake_dnadiff.smk")
     runner.run_workflow(dnadiff_targets_showdiff, config, workdir=Path(tmp_path))
 
     # Check output against target fixtures
@@ -292,9 +290,7 @@ def test_snakemake_rule_show_coords(  # noqa: PLR0913
     # Record the FASTA files in the genomes table _before_ call snakemake
     log_genome(
         database=db,
-        fasta=list(
-            snakemake_scheduler.check_input_stems(config_dnadiff_args["indir"]).values()
-        ),
+        fasta=list(check_input_stems(config_dnadiff_args["indir"]).values()),
     )
     assert db.is_file()
 
@@ -302,7 +298,7 @@ def test_snakemake_rule_show_coords(  # noqa: PLR0913
     config["outdir"] = dnadiff_targets_showcoords_outdir
 
     # Run snakemake wrapper
-    runner = snakemake_scheduler.SnakemakeRunner("snakemake_dnadiff.smk")
+    runner = SnakemakeRunner("snakemake_dnadiff.smk")
     runner.run_workflow(dnadiff_targets_showcoords, config, workdir=Path(tmp_path))
 
     # Check output against target fixtures
