+++ tests/snakemake/test_snakemake_fastani_workflow.py
@@ -34,8 +34,8 @@ from pathlib import Path
 import pytest
 
 from pyani_plus.private_cli import log_configuration, log_genome, log_run
-from pyani_plus.snakemake import snakemake_scheduler
 from pyani_plus.tools import get_fastani
+from pyani_plus.workflows import SnakemakeRunner, check_input_stems
 
 from . import compare_matrices
 
@@ -98,9 +98,7 @@ def test_snakemake_rule_fastani(  # noqa: PLR0913
     # Remove the output directory to force re-running the snakemake rule
     shutil.rmtree(fastani_targets_outdir, ignore_errors=True)
 
-    input_fasta = list(
-        snakemake_scheduler.check_input_stems(config_fastani_args["indir"]).values()
-    )
+    input_fasta = list(check_input_stems(config_fastani_args["indir"]).values())
 
     # Assuming this will match but worker nodes might have a different version
     fastani_tool = get_fastani()
@@ -126,7 +124,7 @@ def test_snakemake_rule_fastani(  # noqa: PLR0913
     assert db.is_file()
 
     # Run snakemake wrapper
-    runner = snakemake_scheduler.SnakemakeRunner("snakemake_fastani.smk")
+    runner = SnakemakeRunner("snakemake_fastani.smk")
 
     runner.run_workflow(fastani_targets, config_fastani_args, workdir=Path(tmp_path))
 
@@ -181,7 +179,7 @@ def test_snakemake_duplicate_stems(
     msg = f"Duplicated stems found for {sorted(stems)}. Please investigate."
 
     # Run snakemake wrapper
-    runner = snakemake_scheduler.SnakemakeRunner("snakemake_fastani.smk")
+    runner = SnakemakeRunner("snakemake_fastani.smk")
 
     with pytest.raises(ValueError, match=re.escape(msg)):
         runner.run_workflow(fastani_targets, dup_config, workdir=Path(tmp_path))
